add test for Environment parameter
